Perl script for running FLASh on pair-ended reads.
This file is a perl script used to run Fast Length Adjustment of Short Reads (FLASh). Here, the script asks the user to provide the files and parameters for running FLASh, and then the Unix program merges the pair-ended reads into single-ended reads. This script is the first to be run and constitutes the first sequence processing step.